Update test check different length than original
Response recorder Header() does not reflect the actual value written to
the network response.

Headers are dumped in the response before the first byte of the body.
httptest.ResponseRecorder implements this behaviour when using the
Result function.

This commit uses this behaviour to ensure checking the actual headers
that would be written in the network response.

As gzip removes the Content-Length on WriteHeader and the gzip encoder
streams the encoded content, there is no possibility to know the actual
response size before writing the headers.

The tests reflects this by asserting Content-Type header is empty

Signed-off-by: Thibault Jamet <tjamet@users.noreply.github.com>